refactor data types from Vec<f32> to Vec<f64> across multiple modules and update related methods

+++ audio_analyzer_core/data/mod.rs
@@ -5,7 +5,7 @@ pub mod test_data;
 pub mod web_stream;
 
 pub trait RawDataStreamLayer {
-    fn try_recv(&mut self) -> Option<Vec<f32>>;
+    fn try_recv(&mut self) -> Option<Vec<f64>>;
     fn sample_rate(&self) -> u32;
     fn start(&mut self);
 }
